Make rounds instantaneous (for the time being)
After thinking long and hard on it I gave up on my initial plan of
making it so each line fades in, stays for a second before fading out
and into the next line.

I want to do that, but I don't know how to
make the rest of the script wait for the text before zooming to the
next line immediately. Limited googling seems to suggest this is not
doable in Javascript, but there's no way, right?

I also learned that innerHTML is perfectly safe as long as you're not
using it to store user inputs.